requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.4
django-crispy-forms==1.11.2
django-heroku==0.3.1
djangorestframework==3.12.4
et-xmlfile==1.0.1
gunicorn==20.1.0
jdcal==1.4.1
openpyxl==3.0.5
Pillow==8.0.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pytz==2020.4
sqlparse==0.4.1
whitenoise==5.3.0
